% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleCount.R
\docType{data}
\name{ExampleCount}
\alias{ExampleCount}
\title{Example Count Data}
\format{
\subsection{\code{ExampleData}}{

A data frame with 33831 rows and 12 columns:
\describe{
\item{symbol}{The HGNC Gene Symbol}
\item{SL31503 to SL7853}{The Gene Counts for each individual sample}
...
}
}
}
\source{
\url{https://www.example.com}
}
\usage{
ExampleCount
}
\description{
The Example Count Data.
From Ramaker Dataset GSE80655,
Region: Anterior Cingulate
Gender: Females
Diagnoses: Bipolar and Control
}
\keyword{datasets}
