% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleMetadata.R
\docType{data}
\name{ExampleMetadata}
\alias{ExampleMetadata}
\title{Example Metadata}
\format{
\subsection{\code{ExampleMetadata}}{

A data frame with 11 rows and 4 columns:
\describe{
\item{Sample-ID}{The Sample ID Corresponding to each column in the count
dataset}
\item{brain_region}{The Brain region the data comes from}
\item{gender}{The Gender of the Sample}
\item{clinical_diagnosis}{The Clinical Diagnosis for the specific Sample.}
}
}
}
\source{
\url{https://www.example.com}
}
\usage{
ExampleMetadata
}
\description{
Accompanying Metadata for the Example Count Data
From Ramaker Dataset GSE80655,
Region: Anterior Cingulate
Gender: Females
Diagnoses: Bipolar and Control
}
\keyword{datasets}
